refactor: rename ContextsStates to ContextsStatesRegistry Signed-off-by: Philippe Martin <phmartin@redhat.com>
