+++ packages/main/src/plugin/kubernetes/contexts-informers-registry.ts
@@ -20,8 +20,8 @@ import type { Informer, KubernetesObject } from '@kubernetes/client-node';
 
 import type { ResourceName } from '/@api/kubernetes-contexts-states.js';
 
-import type { ContextInternalState } from './contexts-states.js';
-import { isSecondaryResourceName } from './contexts-states.js';
+import type { ContextInternalState } from './contexts-states-registry.js';
+import { isSecondaryResourceName } from './contexts-states-registry.js';
 
 export class ContextsInformersRegistry {
   private informers = new Map<string, ContextInternalState>();
