+++ packages/main/src/plugin/kubernetes/contexts-states-registry.spec.ts
@@ -18,7 +18,7 @@
 import { describe, expect, test, vi } from 'vitest';
 
 import type { ApiSenderType } from '../api.js';
-import { ContextsStates, isSecondaryResourceName } from './contexts-states.js';
+import { ContextsStatesRegistry, isSecondaryResourceName } from './contexts-states-registry.js';
 
 const apiSenderSendMock = vi.fn();
 const apiSender: ApiSenderType = {
@@ -33,7 +33,7 @@ describe('ContextsStates tests', () => {
   });
 
   test('isReachable', () => {
-    const client = new ContextsStates(apiSender);
+    const client = new ContextsStatesRegistry(apiSender);
     client.safeSetState('context1', state => (state.reachable = true));
 
     expect(client.isReachable('context1')).toBeTruthy();
@@ -42,7 +42,7 @@ describe('ContextsStates tests', () => {
   });
 
   test('state', () => {
-    const states = new ContextsStates(apiSender);
+    const states = new ContextsStatesRegistry(apiSender);
     expect(states.getContextsGeneralState()).toStrictEqual(new Map());
     expect(states.getContextsCheckingState()).toStrictEqual(new Map());
     expect(states.getCurrentContextGeneralState('ctx1')).toStrictEqual({
